chore: renamed dropdown component not to confuse with dropdown from library
